requirements.txt
flask==3.1.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
werkzeug==3.1.3
sqlalchemy==2.0.36
google-generativeai==0.8.4
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
apscheduler==3.11.0
email-validator==2.2.0
